Move SignUpForm to User module
